Add modular settings table renderer and tests
