+++ src/model/mod.rs
@@ -1,0 +1,6 @@
+//! Types and implementations for machine learning models.
+
+mod preprocessing;
+pub mod supervised;
+
+pub use supervised::SupervisedModel;
