+++ src/settings/settings_struct.rs
@@ -2,20 +2,13 @@
 
 #![allow(clippy::struct_field_names)]
 
-use comfy_table::{
-    Attribute, Cell, Table, modifiers::UTF8_SOLID_INNER_BORDERS, presets::UTF8_FULL,
-};
-
 use super::{
     Algorithm, CategoricalNBParameters, DecisionTreeClassifierParameters,
     DecisionTreeRegressorParameters, ElasticNetParameters, FinalAlgorithm, GaussianNBParameters,
     KNNClassifierParameters, KNNRegressorParameters, LassoParameters, LinearRegressionParameters,
-    LinearRegressionSolverName, LogisticRegressionParameters, Metric, PreProcessing,
-    RandomForestClassifierParameters, RandomForestRegressorParameters, RidgeRegressionParameters,
-    RidgeRegressionSolverName,
+    LogisticRegressionParameters, Metric, PreProcessing, RandomForestClassifierParameters,
+    RandomForestRegressorParameters, RidgeRegressionParameters,
 };
-
-use crate::utils::display::print_option;
 
 use smartcore::{
     metrics::{mean_absolute_error, mean_squared_error, r2},
@@ -29,7 +22,7 @@ use smartcore::linalg::traits::svd::SVDDecomposable;
 use smartcore::numbers::basenum::Number;
 use smartcore::numbers::floatnum::FloatNumber;
 use smartcore::numbers::realnum::RealNumber;
-use std::fmt::{Display, Formatter, Write};
+use std::fmt::{Display, Formatter};
 
 /// Settings for supervised models
 ///
@@ -44,15 +37,15 @@ where
     /// The metric to sort by
     pub(crate) sort_by: Metric,
     /// The type of model to train
-    model_type: ModelType,
+    pub(crate) model_type: ModelType,
     /// The algorithms to skip
     pub(crate) skiplist: Vec<Algorithm<INPUT, OUTPUT, InputArray, OutputArray>>,
     /// The number of folds for cross-validation
-    number_of_folds: usize,
+    pub(crate) number_of_folds: usize,
     /// Whether to shuffle the data
     pub(crate) shuffle: bool,
     /// Whether to be verbose
-    verbose: bool,
+    pub(crate) verbose: bool,
     /// The approach to use for the final model
     pub(crate) final_model_approach: FinalAlgorithm,
     /// The kind of preprocessing to perform
@@ -574,493 +567,9 @@ where
     }
 }
 
-impl<INPUT, OUTPUT, InputArray, OutputArray> Display
-    for Settings<INPUT, OUTPUT, InputArray, OutputArray>
-where
-    INPUT: FloatNumber + RealNumber,
-    OUTPUT: FloatNumber,
-    InputArray: CholeskyDecomposable<INPUT> + SVDDecomposable<INPUT> + QRDecomposable<INPUT>,
-    OutputArray: Array1<OUTPUT>,
-{
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        // Prep new table
-        let mut table = Table::new();
-
-        // Get list of algorithms to skip
-        let mut skiplist = String::new();
-        if self.skiplist.is_empty() {
-            skiplist.push_str("None ");
-        } else {
-            for algorithm_to_skip in &self.skiplist {
-                writeln!(&mut skiplist, "{algorithm_to_skip}")?;
-            }
-        }
-
-        // Build out the table
-        table
-            .load_preset(UTF8_FULL)
-            .apply_modifier(UTF8_SOLID_INNER_BORDERS)
-            .set_header(vec![
-                Cell::new("Settings").add_attribute(Attribute::Bold),
-                Cell::new("Value").add_attribute(Attribute::Bold),
-            ])
-            .add_row(vec![Cell::new("General").add_attribute(Attribute::Italic)])
-            .add_row(vec!["    Model Type", &*format!("{}", self.model_type)])
-            .add_row(vec!["    Verbose", &*format!("{}", self.verbose)])
-            .add_row(vec!["    Sorting Metric", &*format!("{}", self.sort_by)])
-            .add_row(vec!["    Shuffle Data", &*format!("{}", self.shuffle)])
-            .add_row(vec![
-                "    Number of CV Folds",
-                &*format!("{}", self.number_of_folds),
-            ])
-            .add_row(vec![
-                "    Pre-Processing",
-                &*format!("{}", self.preprocessing),
-            ])
-            .add_row(vec![
-                "    Skipped Algorithms",
-                &skiplist[0..skiplist.len() - 1],
-            ]);
-        if !self.skiplist.contains(&Algorithm::default_linear()) {
-            table
-                .add_row(vec![
-                    Cell::new(
-                        Algorithm::<INPUT, OUTPUT, InputArray, OutputArray>::default_linear(),
-                    )
-                    .add_attribute(Attribute::Italic),
-                ])
-                .add_row(vec![
-                    "    Solver",
-                    match self.linear_settings.as_ref().unwrap().solver {
-                        LinearRegressionSolverName::QR => "QR",
-                        LinearRegressionSolverName::SVD => "SVD",
-                    },
-                ]);
-        }
-        if !self.skiplist.contains(&Algorithm::default_ridge()) {
-            table
-                .add_row(vec![
-                    Cell::new(Algorithm::<INPUT, OUTPUT, InputArray, OutputArray>::default_ridge())
-                        .add_attribute(Attribute::Italic),
-                ])
-                .add_row(vec![
-                    "    Solver",
-                    match self.ridge_settings.as_ref().unwrap().solver {
-                        RidgeRegressionSolverName::Cholesky => "Cholesky",
-                        RidgeRegressionSolverName::SVD => "SVD",
-                    },
-                ])
-                .add_row(vec![
-                    "    Alpha",
-                    &*format!("{}", self.ridge_settings.as_ref().unwrap().alpha),
-                ])
-                .add_row(vec![
-                    "    Normalize",
-                    &*format!("{}", self.ridge_settings.as_ref().unwrap().normalize),
-                ]);
-        }
-
-        if !self.skiplist.contains(&Algorithm::default_lasso()) {
-            table
-                .add_row(vec![
-                    Cell::new(Algorithm::<INPUT, OUTPUT, InputArray, OutputArray>::default_lasso())
-                        .add_attribute(Attribute::Italic),
-                ])
-                .add_row(vec![
-                    "    Alpha",
-                    &*format!("{}", self.lasso_settings.as_ref().unwrap().alpha),
-                ])
-                .add_row(vec![
-                    "    Normalize",
-                    &*format!("{}", self.lasso_settings.as_ref().unwrap().normalize),
-                ])
-                .add_row(vec![
-                    "    Maximum Iterations",
-                    &*format!("{}", self.lasso_settings.as_ref().unwrap().max_iter),
-                ])
-                .add_row(vec![
-                    "    Tolerance",
-                    &*format!("{}", self.lasso_settings.as_ref().unwrap().tol),
-                ]);
-        }
-
-        if !self.skiplist.contains(&Algorithm::default_elastic_net()) {
-            table
-                .add_row(vec![
-                    Cell::new(
-                        Algorithm::<INPUT, OUTPUT, InputArray, OutputArray>::default_elastic_net(),
-                    )
-                    .add_attribute(Attribute::Italic),
-                ])
-                .add_row(vec![
-                    "    Alpha",
-                    &*format!("{}", self.elastic_net_settings.as_ref().unwrap().alpha),
-                ])
-                .add_row(vec![
-                    "    Normalize",
-                    &*format!("{}", self.elastic_net_settings.as_ref().unwrap().normalize),
-                ])
-                .add_row(vec![
-                    "    Maximum Iterations",
-                    &*format!("{}", self.elastic_net_settings.as_ref().unwrap().max_iter),
-                ])
-                .add_row(vec![
-                    "    Tolerance",
-                    &*format!("{}", self.elastic_net_settings.as_ref().unwrap().tol),
-                ])
-                .add_row(vec![
-                    "    L1 Ratio",
-                    &*format!("{}", self.elastic_net_settings.as_ref().unwrap().l1_ratio),
-                ]);
-        }
-
-        if !self.skiplist.contains(&Algorithm::default_decision_tree()) {
-            table
-                .add_row(vec![
-                    Cell::new(
-                        Algorithm::<INPUT, OUTPUT, InputArray, OutputArray>::default_decision_tree(
-                        ),
-                    )
-                    .add_attribute(Attribute::Italic),
-                ])
-                .add_row(vec![
-                    "    Max Depth",
-                    &*print_option(
-                        self.decision_tree_regressor_settings
-                            .as_ref()
-                            .unwrap()
-                            .max_depth,
-                    ),
-                ])
-                .add_row(vec![
-                    "    Min samples for leaf",
-                    &*format!(
-                        "{}",
-                        self.decision_tree_regressor_settings
-                            .as_ref()
-                            .unwrap()
-                            .min_samples_leaf
-                    ),
-                ])
-                .add_row(vec![
-                    "    Min samples for split",
-                    &*format!(
-                        "{}",
-                        self.decision_tree_regressor_settings
-                            .as_ref()
-                            .unwrap()
-                            .min_samples_split
-                    ),
-                ]);
-        }
-
-        if !self.skiplist.contains(&Algorithm::default_random_forest()) {
-            table
-                .add_row(vec![
-                    Cell::new(
-                        Algorithm::<INPUT, OUTPUT, InputArray, OutputArray>::default_random_forest(
-                        ),
-                    )
-                    .add_attribute(Attribute::Italic),
-                ])
-                .add_row(vec![
-                    "    Max Depth",
-                    &*print_option(
-                        self.random_forest_regressor_settings
-                            .as_ref()
-                            .unwrap()
-                            .max_depth,
-                    ),
-                ])
-                .add_row(vec![
-                    "    Min samples for leaf",
-                    &*format!(
-                        "{}",
-                        self.random_forest_regressor_settings
-                            .as_ref()
-                            .unwrap()
-                            .min_samples_leaf
-                    ),
-                ])
-                .add_row(vec![
-                    "    Min samples for split",
-                    &*format!(
-                        "{}",
-                        self.random_forest_regressor_settings
-                            .as_ref()
-                            .unwrap()
-                            .min_samples_split
-                    ),
-                ])
-                .add_row(vec![
-                    "    Min samples for split",
-                    &*format!(
-                        "{}",
-                        self.random_forest_regressor_settings
-                            .as_ref()
-                            .unwrap()
-                            .n_trees
-                    ),
-                ])
-                .add_row(vec![
-                    "    Number of split candidates",
-                    &*print_option(self.random_forest_regressor_settings.as_ref().unwrap().m),
-                ]);
-        }
-
-        // if !self.skiplist.contains(&Algorithm::KNNRegressor) {
-        //     table
-        //         .add_row(vec![
-        //             Cell::new(Algorithm::KNNRegressor).add_attribute(Attribute::Italic),
-        //         ])
-        //         .add_row(vec![
-        //             "    Number of neighbors",
-        //             &*format!("{}", self.knn_regressor_settings.as_ref().unwrap().k),
-        //         ])
-        //         .add_row(vec![
-        //             "    Search algorithm",
-        //             &print_knn_search_algorithm(
-        //                 &self.knn_regressor_settings.as_ref().unwrap().algorithm,
-        //             ),
-        //         ])
-        //         .add_row(vec![
-        //             "    Weighting function",
-        //             &print_knn_weight_function(
-        //                 &self.knn_regressor_settings.as_ref().unwrap().weight,
-        //             ),
-        //         ])
-        //         .add_row(vec![
-        //             "    Distance function",
-        //             &*format!(
-        //                 "{}",
-        //                 &self.knn_regressor_settings.as_ref().unwrap().distance
-        //             ),
-        //         ]);
-        // }
-
-        // if !self.skiplist.contains(&Algorithm::SVR) {
-        //     table
-        //         .add_row(vec![
-        //             Cell::new(Algorithm::SVR).add_attribute(Attribute::Italic),
-        //         ])
-        //         .add_row(vec![
-        //             "    Regularization parameter",
-        //             &*format!("{}", self.svr_settings.as_ref().unwrap().c),
-        //         ])
-        //         .add_row(vec![
-        //             "    Tolerance",
-        //             &*format!("{}", self.svr_settings.as_ref().unwrap().tol),
-        //         ])
-        //         .add_row(vec![
-        //             "    Epsilon",
-        //             &*format!("{}", self.svr_settings.as_ref().unwrap().eps),
-        //         ])
-        //         .add_row(vec![
-        //             "    Kernel",
-        //             &*format!("{}", self.svr_settings.as_ref().unwrap().kernel),
-        //         ]);
-        // }
-
-        // if !self.skiplist.contains(&Algorithm::LogisticRegression) {
-        //     table
-        //         .add_row(vec![
-        //             Cell::new(Algorithm::LogisticRegression).add_attribute(Attribute::Italic),
-        //         ])
-        //         .add_row(vec!["    N/A", "N/A"]);
-        // }
-
-        // if !self.skiplist.contains(&Algorithm::default_random_forest()) {
-        //     table
-        //         .add_row(vec![Cell::new(Algorithm::<INPUT, OUTPUT, InputArray, OutputArray>::default_random_forest()).add_attribute(Attribute::Italic)])
-        //         .add_row(vec![
-        //             "    Split Criterion",
-        //             match self
-        //                 .random_forest_classifier_settings
-        //                 .as_ref()
-        //                 .unwrap()
-        //                 .criterion
-        //             {
-        //                 SplitCriterion::Gini => "Gini",
-        //                 SplitCriterion::Entropy => "Entropy",
-        //                 SplitCriterion::ClassificationError => "Classification Error",
-        //             },
-        //         ])
-        //         .add_row(vec![
-        //             "    Max Depth",
-        //             &*print_option(
-        //                 self.random_forest_classifier_settings
-        //                     .as_ref()
-        //                     .unwrap()
-        //                     .max_depth,
-        //             ),
-        //         ])
-        //         .add_row(vec![
-        //             "    Min samples for leaf",
-        //             &*format!(
-        //                 "{}",
-        //                 self.random_forest_classifier_settings
-        //                     .as_ref()
-        //                     .unwrap()
-        //                     .min_samples_leaf
-        //             ),
-        //         ])
-        //         .add_row(vec![
-        //             "    Min samples for split",
-        //             &*format!(
-        //                 "{}",
-        //                 self.random_forest_classifier_settings
-        //                     .as_ref()
-        //                     .unwrap()
-        //                     .min_samples_split
-        //             ),
-        //         ])
-        //         .add_row(vec![
-        //             "    Min samples for split",
-        //             &*format!(
-        //                 "{}",
-        //                 self.random_forest_classifier_settings
-        //                     .as_ref()
-        //                     .unwrap()
-        //                     .n_trees
-        //             ),
-        //         ])
-        //         .add_row(vec![
-        //             "    Number of split candidates",
-        //             &*print_option(self.random_forest_classifier_settings.as_ref().unwrap().m),
-        //         ]);
-        // }
-        //
-        // if !self.skiplist.contains(&Algorithm::KNNClassifier) {
-        //     table
-        //         .add_row(vec![
-        //             Cell::new(Algorithm::KNNClassifier).add_attribute(Attribute::Italic),
-        //         ])
-        //         .add_row(vec![
-        //             "    Number of neighbors",
-        //             &*format!("{}", self.knn_classifier_settings.as_ref().unwrap().k),
-        //         ])
-        //         .add_row(vec![
-        //             "    Search algorithm",
-        //             &print_knn_search_algorithm(
-        //                 &self.knn_classifier_settings.as_ref().unwrap().algorithm,
-        //             ),
-        //         ])
-        //         .add_row(vec![
-        //             "    Weighting function",
-        //             &print_knn_weight_function(
-        //                 &self.knn_classifier_settings.as_ref().unwrap().weight,
-        //             ),
-        //         ])
-        //         .add_row(vec![
-        //             "    Distance function",
-        //             &*format!(
-        //                 "{}",
-        //                 &self.knn_classifier_settings.as_ref().unwrap().distance
-        //             ),
-        //         ]);
-        // }
-        //
-        // if !self.skiplist.contains(&Algorithm::SVC) {
-        //     table
-        //         .add_row(vec![
-        //             Cell::new(Algorithm::SVC).add_attribute(Attribute::Italic),
-        //         ])
-        //         .add_row(vec![
-        //             "    Regularization parameter",
-        //             &*format!("{}", self.svc_settings.as_ref().unwrap().c),
-        //         ])
-        //         .add_row(vec![
-        //             "    Tolerance",
-        //             &*format!("{}", self.svc_settings.as_ref().unwrap().tol),
-        //         ])
-        //         .add_row(vec![
-        //             "    Epoch",
-        //             &*format!("{}", self.svc_settings.as_ref().unwrap().epoch),
-        //         ])
-        //         .add_row(vec![
-        //             "    Kernel",
-        //             &*format!("{}", self.svc_settings.as_ref().unwrap().kernel),
-        //         ]);
-        // }
-        //
-        // if !self.skiplist.contains(&Algorithm::DecisionTreeClassifier) {
-        //     table
-        //         .add_row(vec![
-        //             "    Split Criterion",
-        //             match self
-        //                 .random_forest_classifier_settings
-        //                 .as_ref()
-        //                 .unwrap()
-        //                 .criterion
-        //             {
-        //                 SplitCriterion::Gini => "Gini",
-        //                 SplitCriterion::Entropy => "Entropy",
-        //                 SplitCriterion::ClassificationError => "Classification Error",
-        //             },
-        //         ])
-        //         .add_row(vec![
-        //             Cell::new(Algorithm::DecisionTreeClassifier).add_attribute(Attribute::Italic),
-        //         ])
-        //         .add_row(vec![
-        //             "    Max Depth",
-        //             &*print_option(
-        //                 self.decision_tree_classifier_settings
-        //                     .as_ref()
-        //                     .unwrap()
-        //                     .max_depth,
-        //             ),
-        //         ])
-        //         .add_row(vec![
-        //             "    Min samples for leaf",
-        //             &*format!(
-        //                 "{}",
-        //                 self.decision_tree_classifier_settings
-        //                     .as_ref()
-        //                     .unwrap()
-        //                     .min_samples_leaf
-        //             ),
-        //         ])
-        //         .add_row(vec![
-        //             "    Min samples for split",
-        //             &*format!(
-        //                 "{}",
-        //                 self.decision_tree_classifier_settings
-        //                     .as_ref()
-        //                     .unwrap()
-        //                     .min_samples_split
-        //             ),
-        //         ]);
-        // }
-        //
-        // if !self.skiplist.contains(&Algorithm::CategoricalNaiveBayes) {
-        //     table
-        //         .add_row(vec![
-        //             Cell::new(Algorithm::CategoricalNaiveBayes).add_attribute(Attribute::Italic),
-        //         ])
-        //         .add_row(vec![
-        //             "    Smoothing parameter",
-        //             &*format!("{}", self.categorical_nb_settings.as_ref().unwrap().alpha),
-        //         ]);
-        // }
-        //
-        // if !self.skiplist.contains(&Algorithm::GaussianNaiveBayes) {
-        //     table
-        //         .add_row(vec![
-        //             Cell::new(Algorithm::GaussianNaiveBayes).add_attribute(Attribute::Italic),
-        //         ])
-        //         .add_row(vec![
-        //             "    Priors",
-        //             &*debug_option(self.gaussian_nb_settings.as_ref().unwrap().clone().priors),
-        //         ]);
-        // }
-
-        writeln!(f, "{table}")
-    }
-}
-
 /// Model type to train
 #[derive(serde::Serialize, serde::Deserialize)]
-enum ModelType {
+pub(crate) enum ModelType {
     /// No model type specified
     None,
     /// Regression model
